pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^21.1.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
